refactor(daemon): move DaemonLaunchDescriptor from :mcp into the published daemon-core contract
Everything that launches a preview daemon needs the `daemon-launch.json` reader — the MCP
supervisor, `:render-session-subprocess`, and `compose-preview serve`'s bundle daemon. It lived
inside `:mcp`, so every one of them was a consumer of the MCP server module.

That is the coupling issue #3824 has to remove before `serve` can be extracted: the preview server
is a protocol client, and its dependency floor should be `daemon-core` plus the payload schemas, not
an MCP server. The descriptor is a self-contained @Serializable data class with no MCP in it; moving
it to `:daemon:core` puts the reader on the right side of the boundary and drops serve's only direct
`ee.schimke.composeai.mcp` import.

This is half of preparation item 2. The other half remains: this reader and the gradle plugin's
published `daemon-launch-builder` writer are two representations of one schema and have drifted.

`:mcp` still reaches the classpath of anything using `:render-session-subprocess`, because that
module builds its transport on `:mcp`'s `DaemonClient` — preparation item 3, recorded as a known
leak in the contract probe.
