+++ mcp/src/main/kotlin/ee/schimke/composeai/mcp/DaemonSupervisor.kt
@@ -1,12 +1,12 @@
 package ee.schimke.composeai.mcp
 
+import ee.schimke.composeai.daemon.DaemonLaunchDescriptor
 import ee.schimke.composeai.daemon.protocol.BackendKind
 import ee.schimke.composeai.daemon.protocol.DataProductCapability
 import ee.schimke.composeai.io.SystemFileSystem
 import java.io.File
 import java.util.concurrent.ConcurrentHashMap
 import kotlin.time.Duration
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.contentOrNull
@@ -680,82 +680,6 @@ fun interface DescriptorProvider {
         }
       return index
     }
-  }
-}
-
-/**
- * Trimmed parse of `build/compose-previews/daemon-launch.json`. Mirrors the field set written by
- * [`DaemonClasspathDescriptor`][ee.schimke.composeai.plugin.daemon.DaemonClasspathDescriptor] in
- * the gradle plugin; we re-declare the schema rather than depending on the plugin module so the MCP
- * module's runtime classpath stays free of the plugin's AGP/Gradle deps.
- */
-@Serializable
-data class DaemonLaunchDescriptor(
-  val schemaVersion: Int,
-  val modulePath: String,
-  val variant: String,
-  val enabled: Boolean,
-  val mainClass: String,
-  val javaLauncher: String? = null,
-  val classpath: List<String>,
-  val jvmArgs: List<String>,
-  val systemProperties: Map<String, String>,
-  val workingDirectory: String,
-  val manifestPath: String,
-  /**
-   * Optional argv **prefix** the daemon JVM launches behind — an OS jail (`bwrap`, `unshare`,
-   * `systemd-run --scope`, …). Empty (the default, and what the gradle plugin writes) launches the
-   * JVM directly, exactly as before.
-   *
-   * This is how the playground's per-session sandbox reaches the live lane: serve writes the jail
-   * into the snippet's own `daemon-launch.json`, so the descriptor→spawn path applies it without
-   * every intermediate layer having to thread a sandbox object through (`docs/design/PLAYGROUND.md`
-   * §6). Distinct from [withSandboxCount], which sizes Robolectric's *in-JVM* sandbox pool and has
-   * nothing to do with containment.
-   */
-  val jailCommand: List<String> = emptyList(),
-  /**
-   * Optional hard wall-clock lifetime, in seconds, for the spawned JVM. When set, the spawner arms
-   * a watchdog that force-kills the process at the deadline — the "killed after a hard wall-clock
-   * TTL" requirement for a playground session, enforced by the parent rather than by the child's
-   * cooperation. Null (the default) means no watchdog: an ordinary project daemon lives as long as
-   * its owner keeps it.
-   */
-  val hardTtlSeconds: Long? = null,
-) {
-
-  /** Returns a copy launched behind [command] and force-killed after [hardTtlSeconds]. */
-  fun jailed(command: List<String>, hardTtlSeconds: Long?): DaemonLaunchDescriptor =
-    copy(jailCommand = command, hardTtlSeconds = hardTtlSeconds)
-
-  /**
-   * SANDBOX-POOL.md — returns a copy with `composeai.daemon.sandboxCount` merged into
-   * [systemProperties]. The supervisor calls this on the descriptor read from disk before passing
-   * it to [DaemonClientFactory.spawn] so the daemon JVM picks up the right pool size at boot.
-   *
-   * Idempotent at [count] = 1 (the daemon's default; the sysprop is omitted to keep the disk
-   * descriptor trivially diffable across replicas-per-daemon settings of 0).
-   */
-  fun withSandboxCount(count: Int): DaemonLaunchDescriptor {
-    require(count >= 1) { "sandboxCount must be >= 1, got $count" }
-    if (count == 1) return this
-    val merged = systemProperties.toMutableMap()
-    merged[SANDBOX_COUNT_PROP] = count.toString()
-    return copy(systemProperties = merged)
-  }
-
-  companion object {
-    /**
-     * Sysprop key the daemon reads to configure
-     * [`RobolectricHost.sandboxCount`][ee.schimke.composeai.daemon.RobolectricHost.sandboxCount].
-     * Mirrored on the daemon side as a private const in `DaemonMain.kt`; both sides MUST agree.
-     */
-    const val SANDBOX_COUNT_PROP: String = "composeai.daemon.sandboxCount"
-
-    private val json = Json { ignoreUnknownKeys = true }
-
-    fun parse(jsonText: String): DaemonLaunchDescriptor =
-      json.decodeFromString(serializer(), jsonText)
   }
 }
